requirements.txt
pillow
pygame
pandas
asyncio
edge_tts
openpyxl